Move wallet files and adjust imports
