+++ fe1-web/features/rollCall/network/__mocks__/RollCallMessageApi.ts
@@ -1,4 +1,4 @@
 /**
- * Mocks functions to send messages roll call messages.
+ * Mocks functions to send roll call messages.
  */
 export const requestCloseRollCall = jest.fn(() => Promise.resolve());
