+++ fe1-web/features/wallet/navigation/WalletNavigation.tsx
@@ -3,11 +3,13 @@ import { createStackNavigator } from '@react-navigation/stack';
 
 import STRINGS from 'res/strings';
 
-import WalletHome from 'parts/wallet/WalletHome';
-import WalletSetSeed from 'parts/wallet/WalletSetSeed';
-import WalletShowSeed from 'parts/wallet/WalletShowSeed';
-import WalletSyncedSeed from 'parts/wallet/WalletSyncedSeed';
-import WalletError from 'parts/wallet/WalletError';
+import {
+  WalletError,
+  WalletHome,
+  WalletSetSeed,
+  WalletShowSeed,
+  WalletSyncedSeed,
+} from '../screens';
 
 /**
  * Define the Wallet stack navigation
