+++ fe1-web/features/wallet/objects/index.ts
@@ -1,0 +1,3 @@
+export * from './Wallet';
+export * from './Seed';
+export { generateToken, getCurrentPopTokenFromStore } from './Token';
