+++ fe1-web/features/wallet/screens/WalletSyncedSeed.tsx
@@ -1,7 +1,6 @@
 import React, { ReactNode, useEffect, useState } from 'react';
 import {
-  ScrollView,
-  StyleSheet, View, ViewStyle,
+  ScrollView, StyleSheet, View, ViewStyle,
 } from 'react-native';
 import PropTypes from 'prop-types';
 import { useSelector } from 'react-redux';
@@ -10,7 +9,7 @@ import containerStyles from 'styles/stylesheets/containerStyles';
 import STRINGS from 'res/strings';
 import TextBlock from 'components/TextBlock';
 import WideButtonView from 'components/WideButtonView';
-import { PopToken, Wallet } from 'model/objects';
+import { PopToken } from 'model/objects';
 import QRCode from 'components/QRCode';
 import PROPS_TYPE from 'res/Props';
 import { makeLaosMap } from 'store';
@@ -18,6 +17,8 @@ import CopiableTextInput from 'components/CopiableTextInput';
 import { makeEventByTypeSelector } from 'features/events/reducer/EventsReducer';
 import { LaoEventType } from 'features/events/objects';
 import { RollCall } from 'features/rollCall/objects';
+
+import * as Wallet from '../objects';
 
 const styles = StyleSheet.create({
   smallPadding: {
