+++ fe1-web/features/wallet/screens/index.ts
@@ -1,0 +1,5 @@
+export { default as WalletError } from './WalletError';
+export { default as WalletHome } from './WalletHome';
+export { default as WalletSetSeed } from './WalletSetSeed';
+export { default as WalletShowSeed } from './WalletShowSeed';
+export { default as WalletSyncedSeed } from './WalletSyncedSeed';
