+++ fe1-web/features/wallet/store/index.ts
@@ -1,0 +1,1 @@
+export * from './WalletStore';
