+++ fe1-web/model/objects/index.ts
@@ -1,7 +1,3 @@
-import * as w from './wallet';
-
-export const Wallet = w;
-
 export * from './Base64Url';
 export * from './Channel';
 export * from './ConnectToLao';
