--- fe1-web/model/objects/wallet/index.ts
@@ -1,6 +1,0 @@
-export * from './Wallet';
-export * from './Seed';
-export {
-  generateToken,
-  getCurrentPopTokenFromStore,
-} from './Token';
