+++ fe1-web/store/__mocks__/Storage.ts
@@ -2,7 +2,7 @@ import {
   applyMiddleware, Store, createStore, combineReducers,
 } from 'redux';
 import thunkMiddleware from 'redux-thunk';
-import walletReducer from 'store/reducers/WalletReducer';
+import walletReducer from 'features/wallet/reducer/WalletReducer';
 
 const reducers = combineReducers({
   ...walletReducer,
