+++ fe1-web/store/reducers/RootReducer.ts
@@ -3,10 +3,10 @@ import { persistCombineReducers } from 'redux-persist';
 
 import socialReducer from 'features/social/reducer/SocialReducer';
 import eventsReducer from 'features/events/reducer/EventsReducer';
+import walletReducer from 'features/wallet/reducer/WalletReducer';
 import laoReducer from './LaoReducer';
 import keyPairReducer from './KeyPairReducer';
 import messageReducer from './MessageReducer';
-import walletReducer from './WalletReducer';
 
 import { wrapWithClearStorageReducer } from './ClearStorageReducer';
