+++ fe1-web/store/reducers/index.ts
@@ -22,4 +22,3 @@ export default rootReducer;
 export * from './LaoReducer';
 export * from './KeyPairReducer';
 export * from './MessageReducer';
-export * from './WalletReducer';
