+++ fe1-web/store/stores/index.ts
@@ -11,4 +11,3 @@
 
 export * from './KeyPairStore';
 export * from './OpenedLaoStore';
-export * from './WalletStore';
